Add file management actions to the UI
* Add Save action
* Add Exit action
* Add Open action
* Add New Action

+++ src/fll/xml/ui/ChallengeDescriptionEditor.java
@@ -16,10 +16,12 @@ import java.awt.event.KeyEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.Reader;
 import java.io.Writer;
 import java.util.LinkedList;
 import java.util.List;
@@ -32,6 +34,8 @@ import javax.swing.Box;
 import javax.swing.JComponent;
 import javax.swing.JFileChooser;
 import javax.swing.JFrame;
+import javax.swing.JMenu;
+import javax.swing.JMenuBar;
 import javax.swing.JOptionPane;
 import javax.swing.JScrollPane;
 import javax.swing.JToolBar;
@@ -50,6 +54,7 @@ import fll.util.GuiExceptionHandler;
 import fll.util.LogUtils;
 import fll.xml.ChallengeDescription;
 import fll.xml.ChallengeParser;
+import fll.xml.ChallengeXMLException;
 import fll.xml.SubjectiveScoreCategory;
 import fll.xml.XMLUtils;
 import fll.xml.ui.MovableExpandablePanel.MoveEvent;
@@ -60,8 +65,10 @@ import net.mtu.eggplant.util.gui.GraphicsUtils;
 
 /*
  * TODO
- * * Be able to create all elements without an XML element
- * - add validity check somewhere to ensure that all names are unique, perhaps this is part of the UI
+ * - add validity check 
+ *   - call from the UI
+ *     - display indicator on each element if it's valid or not?
+ *   - ensure that all names are unique
  *   - see the validity check that is run after parsing the XML
  *   - non-null name for AbstractGoal title and name
  *   - SwitchStatement must have a default case
@@ -73,16 +80,16 @@ import net.mtu.eggplant.util.gui.GraphicsUtils;
  *   - RubricRange min/max
  *   - unique variable names inside ComputedGoal
  *   - SwitchStatement must have something in the default case
- * - choose challenge description to edit
- * - support grouping of goals, this is where DnD might be useful 
- *   - all goals in a group must be consecutive
- * - how to handle when being run from the launcher so that it doesn't exit?
- *   - maybe have a method that is called from main that creates everything
- * - track the name of the file and support save functionality
+ * - ability to add subjective categories
+ * - Ability to add goals
  * - Note that one can create a BasicPolynomial with variable references.
  *   - Perhaps BasicPolynomial and ComplexPolynomial can be merged?
  *   - What effect does this have on the XML? Is there always a scope available?
  *   - If we can't do this, then the UI needs to enforce the restriction.
+ * - support grouping of goals, this is where DnD might be useful 
+ *   - all goals in a group must be consecutive
+ * - how to handle when being run from the launcher so that it doesn't exit?
+ *   - maybe have a method that is called from main that creates everything
  * 
  */
 
@@ -124,7 +131,8 @@ public class ChallengeDescriptionEditor extends JFrame {
 
       final ChallengeDescription description = new ChallengeDescription(challengeDocument.getDocumentElement());
 
-      final ChallengeDescriptionEditor editor = new ChallengeDescriptionEditor(description);
+      final ChallengeDescriptionEditor editor = new ChallengeDescriptionEditor();
+      editor.setChallengeDescription(null, description);
 
       // FIXME need close handler, think about running from launcher and how
       // that will work
@@ -155,15 +163,24 @@ public class ChallengeDescriptionEditor extends JFrame {
     }
   }
 
-  private final ChallengeDescription mDescription;
+  private ChallengeDescription mDescription;
+
+  private File mCurrentFile = null;
 
   private final ScoreCategoryEditor mPerformanceEditor;
 
   private final List<ScoreCategoryEditor> mSubjectiveEditors = new LinkedList<>();
 
-  public ChallengeDescriptionEditor(final ChallengeDescription description) {
+  private final JComponent mSubjectiveContainer;
+
+  private final MoveEventListener mSubjectiveMoveListener;
+
+  public ChallengeDescriptionEditor() {
     super("Challenge Description Editor");
-    mDescription = description;
+    mDescription = null;
+    mCurrentFile = null;
+
+    createMenuBar();
 
     final Container cpane = getContentPane();
     cpane.setLayout(new BorderLayout());
@@ -173,24 +190,24 @@ public class ChallengeDescriptionEditor extends JFrame {
     final JComponent topPanel = Box.createVerticalBox();
     topPanel.setAlignmentX(LEFT_ALIGNMENT);
 
-    mPerformanceEditor = new ScoreCategoryEditor(mDescription.getPerformance());
+    mPerformanceEditor = new ScoreCategoryEditor();
     mPerformanceEditor.setBorder(BorderFactory.createEtchedBorder(EtchedBorder.RAISED));
     final MovableExpandablePanel performance = new MovableExpandablePanel("Performance", mPerformanceEditor, false);
     topPanel.add(performance);
 
-    final JComponent subjective = Box.createVerticalBox();
-    subjective.setBorder(BorderFactory.createTitledBorder("Subjective"));
-    topPanel.add(subjective);
+    mSubjectiveContainer = Box.createVerticalBox();
+    mSubjectiveContainer.setBorder(BorderFactory.createTitledBorder("Subjective"));
+    topPanel.add(mSubjectiveContainer);
 
-    final MoveEventListener subjectiveMoveListener = new MoveEventListener() {
+    mSubjectiveMoveListener = new MoveEventListener() {
 
       @Override
       public void requestedMove(final MoveEvent e) {
         // find index of e.component in subjective
         int oldIndex = -1;
         for (int i = 0; oldIndex < 0
-            && i < subjective.getComponentCount(); ++i) {
-          final Component c = subjective.getComponent(i);
+            && i < mSubjectiveContainer.getComponentCount(); ++i) {
+          final Component c = mSubjectiveContainer.getComponent(i);
           if (c == e.getComponent()) {
             oldIndex = i;
           }
@@ -212,7 +229,7 @@ public class ChallengeDescriptionEditor extends JFrame {
         }
 
         if (newIndex < 0
-            || newIndex >= subjective.getComponentCount()) {
+            || newIndex >= mSubjectiveContainer.getComponentCount()) {
           if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Can't move component outside the container oldIndex: "
                 + oldIndex
@@ -223,8 +240,8 @@ public class ChallengeDescriptionEditor extends JFrame {
         }
 
         // update the UI
-        subjective.add(e.getComponent(), newIndex);
-        subjective.validate();
+        mSubjectiveContainer.add(e.getComponent(), newIndex);
+        mSubjectiveContainer.validate();
 
         // update the order in the challenge description
         final SubjectiveScoreCategory category = mDescription.removeSubjectiveCategory(oldIndex);
@@ -233,28 +250,207 @@ public class ChallengeDescriptionEditor extends JFrame {
 
     };
 
-    for (final SubjectiveScoreCategory cat : mDescription.getSubjectiveCategories()) {
-      final ScoreCategoryEditor editor = new ScoreCategoryEditor(cat);
-      editor.setBorder(BorderFactory.createEtchedBorder(EtchedBorder.RAISED));
-
-      final MovableExpandablePanel container = new MovableExpandablePanel(cat.getTitle(), editor);
-      container.addMoveEventListener(subjectiveMoveListener);
-
-      subjective.add(container);
-
-      mSubjectiveEditors.add(editor);
-    }
-
     // fill in the bottom of the panel
     topPanel.add(Box.createVerticalGlue());
 
     final JScrollPane scroller = new JScrollPane(topPanel);
     cpane.add(scroller, BorderLayout.CENTER);
 
-    pack();
+    setSize(800, 600);
   }
 
-  private final Action mSaveAction = new AbstractAction("Save As...") {
+  /**
+   * Create the menu bar for the frame and set it on the frame.
+   */
+  private void createMenuBar() {
+    final JMenuBar menubar = new JMenuBar();
+
+    menubar.add(createFileMenu());
+
+    this.setJMenuBar(menubar);
+  }
+
+  /**
+   * @return the file menu
+   */
+  private JMenu createFileMenu() {
+    final JMenu menu = new JMenu("File");
+    menu.setMnemonic(KeyEvent.VK_F);
+
+    menu.add(mNewAction);
+    menu.add(mOpenAction);
+
+    menu.addSeparator();
+
+    menu.add(mSaveAction);
+    menu.add(mSaveAsAction);
+
+    menu.addSeparator();
+
+    menu.add(mExitAction);
+
+    return menu;
+  }
+
+  private final Action mOpenAction = new AbstractAction("Open...") {
+    {
+      putValue(SMALL_ICON, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Open16.gif"));
+      putValue(LARGE_ICON_KEY, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Open24.gif"));
+      putValue(SHORT_DESCRIPTION, "Open an existing challenge description");
+      putValue(MNEMONIC_KEY, KeyEvent.VK_O);
+      putValue(ACCELERATOR_KEY, KeyStroke.getKeyStroke(KeyEvent.VK_O, ActionEvent.CTRL_MASK));
+    }
+
+    @Override
+    public void actionPerformed(final ActionEvent ae) {
+      // FIXME prompt for saving of existing file
+
+      openChallengeDescription();
+    }
+  };
+
+  private final Action mNewAction = new AbstractAction("New") {
+    {
+      putValue(SMALL_ICON, GraphicsUtils.getIcon("toolbarButtonGraphics/general/New16.gif"));
+      putValue(LARGE_ICON_KEY, GraphicsUtils.getIcon("toolbarButtonGraphics/general/New24.gif"));
+      putValue(SHORT_DESCRIPTION, "Create a new challenge description");
+      putValue(MNEMONIC_KEY, KeyEvent.VK_N);
+      putValue(ACCELERATOR_KEY, KeyStroke.getKeyStroke(KeyEvent.VK_N, ActionEvent.CTRL_MASK));
+    }
+
+    @Override
+    public void actionPerformed(final ActionEvent ae) {
+      // FIXME prompt for saving of existing file
+
+      setChallengeDescription(null, null);
+    }
+  };
+
+  private void openChallengeDescription() {
+    final String startingDirectory = PREFS.get(DESCRIPTION_STARTING_DIRECTORY_PREF, null);
+
+    final JFileChooser fileChooser = new JFileChooser();
+    final FileFilter filter = new BasicFileFilter("FLL Challenge Description (xml)", new String[] { "xml" });
+    fileChooser.setFileFilter(filter);
+    if (null != startingDirectory) {
+      fileChooser.setCurrentDirectory(new File(startingDirectory));
+    }
+
+    final int returnVal = fileChooser.showOpenDialog(this);
+    if (returnVal == JFileChooser.APPROVE_OPTION) {
+      final File currentDirectory = fileChooser.getCurrentDirectory();
+      PREFS.put(DESCRIPTION_STARTING_DIRECTORY_PREF, currentDirectory.getAbsolutePath());
+
+      final File file = fileChooser.getSelectedFile();
+
+      if (!file.exists()) {
+        if (LOGGER.isTraceEnabled()) {
+          LOGGER.trace("User asked to open a file that doesn't exist");
+        }
+        JOptionPane.showMessageDialog(this, String.format("The file %s doesn't exist", file.getAbsolutePath()), "Error",
+                                      JOptionPane.ERROR_MESSAGE);
+        return;
+      }
+
+      try (Reader stream = new FileReader(file)) {
+
+        final Document challengeDocument = ChallengeParser.parse(stream);
+
+        final ChallengeDescription description = new ChallengeDescription(challengeDocument.getDocumentElement());
+
+        setChallengeDescription(file, description);
+      } catch (final IOException e) {
+        LOGGER.error("Error loading "
+            + file.getAbsolutePath(), e);
+
+        JOptionPane.showMessageDialog(null, "Error reading the XML file. Error message: "
+            + e.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
+
+      } catch (final ChallengeXMLException e) {
+        LOGGER.error("Malformed XML "
+            + file.getAbsolutePath(), e);
+
+        JOptionPane.showMessageDialog(null, "The XML file has an error in it. Error message: "
+            + e.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
+
+      }
+    }
+  }
+
+  /**
+   * @param file the file that the challenge was loaded from, may be null
+   * @param description the description to edit, if null create a new challenge
+   *          description
+   */
+  private void setChallengeDescription(final File file,
+                                       final ChallengeDescription description) {
+    mCurrentFile = file;
+    mDescription = description == null ? new ChallengeDescription("New Challenge") : description;
+
+    if (null != mCurrentFile) {
+      setTitle(mCurrentFile.getName());
+    }
+
+    mSubjectiveContainer.removeAll();
+    mSubjectiveEditors.clear();
+
+    mPerformanceEditor.setCategory(mDescription.getPerformance());
+
+    for (final SubjectiveScoreCategory cat : mDescription.getSubjectiveCategories()) {
+      final ScoreCategoryEditor editor = new ScoreCategoryEditor();
+      editor.setCategory(cat);
+      editor.setBorder(BorderFactory.createEtchedBorder(EtchedBorder.RAISED));
+
+      final MovableExpandablePanel container = new MovableExpandablePanel(cat.getTitle(), editor);
+      container.addMoveEventListener(mSubjectiveMoveListener);
+
+      mSubjectiveContainer.add(container);
+
+      mSubjectiveEditors.add(editor);
+    }
+    
+    validate();
+  }
+
+  private final Action mExitAction = new AbstractAction("Exit") {
+    {
+      putValue(SMALL_ICON, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Stop16.gif"));
+      putValue(LARGE_ICON_KEY, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Stop24.gif"));
+      putValue(SHORT_DESCRIPTION, "Exit the application");
+      putValue(MNEMONIC_KEY, KeyEvent.VK_Q);
+      putValue(ACCELERATOR_KEY, KeyStroke.getKeyStroke(KeyEvent.VK_Q, ActionEvent.CTRL_MASK));
+    }
+
+    @Override
+    public void actionPerformed(final ActionEvent ae) {
+      exit();
+    }
+  };
+
+  private void exit() {
+    // FIXME: check that everything is saved
+
+    // only hide the frame so that when being called from elsewhere the system
+    // doesn't exit
+    setVisible(false);
+  }
+
+  private final Action mSaveAction = new AbstractAction("Save") {
+    {
+      putValue(SMALL_ICON, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Save16.gif"));
+      putValue(LARGE_ICON_KEY, GraphicsUtils.getIcon("toolbarButtonGraphics/general/Save24.gif"));
+      putValue(SHORT_DESCRIPTION, "Save the challenge description file with the current name");
+      putValue(MNEMONIC_KEY, KeyEvent.VK_S);
+      putValue(ACCELERATOR_KEY, KeyStroke.getKeyStroke(KeyEvent.VK_S, ActionEvent.CTRL_MASK));
+    }
+
+    @Override
+    public void actionPerformed(final ActionEvent ae) {
+      saveChallengeDescription();
+    }
+  };
+
+  private final Action mSaveAsAction = new AbstractAction("Save As...") {
     {
       putValue(SMALL_ICON, GraphicsUtils.getIcon("toolbarButtonGraphics/general/SaveAs16.gif"));
       putValue(LARGE_ICON_KEY, GraphicsUtils.getIcon("toolbarButtonGraphics/general/SaveAs24.gif"));
@@ -266,7 +462,7 @@ public class ChallengeDescriptionEditor extends JFrame {
 
     @Override
     public void actionPerformed(final ActionEvent ae) {
-      saveChallengeDescription();
+      saveAsChallengeDescription();
     }
   };
 
@@ -274,7 +470,35 @@ public class ChallengeDescriptionEditor extends JFrame {
 
   private static final String DESCRIPTION_STARTING_DIRECTORY_PREF = "descriptionStartingDirectory";
 
+  /**
+   * Save the challenge description to the current filename. If the current file
+   * is null, then call {@link #saveAsChallengeDescription()}.
+   */
   private void saveChallengeDescription() {
+    if (null == mCurrentFile) {
+      saveAsChallengeDescription();
+      return;
+    }
+
+    commitChanges();
+
+    try (final Writer writer = new FileWriter(mCurrentFile)) {
+      final Document saveDoc = mDescription.toXml();
+      XMLUtils.writeXML(saveDoc, writer, Utilities.DEFAULT_CHARSET.name());
+    } catch (final IOException e) {
+      LOGGER.error("Error writing document", e);
+
+      JOptionPane.showMessageDialog(null,
+                                    "An unexpected error occurred. Please send the log file and a description of what you were doing to the developer. Error message: "
+                                        + e.getMessage(),
+                                    "Error", JOptionPane.ERROR_MESSAGE);
+    }
+  }
+
+  /**
+   * Save the challenge description and prompt the user for the filename.
+   */
+  private void saveAsChallengeDescription() {
     commitChanges();
 
     final String startingDirectory = PREFS.get(DESCRIPTION_STARTING_DIRECTORY_PREF, null);
@@ -297,9 +521,21 @@ public class ChallengeDescriptionEditor extends JFrame {
             + ".xml");
       }
 
+      if (file.exists()) {
+        final int result = JOptionPane.showConfirmDialog(this,
+                                                         String.format("The file %s already exists, do you want to overwrite it?",
+                                                                       file.getName()),
+                                                         "Overwrite file?", JOptionPane.YES_NO_OPTION);
+        if (result == JOptionPane.NO_OPTION) {
+          return;
+        }
+      }
+
       try (final Writer writer = new FileWriter(file)) {
         final Document saveDoc = mDescription.toXml();
         XMLUtils.writeXML(saveDoc, writer, Utilities.DEFAULT_CHARSET.name());
+
+        mCurrentFile = file;
       } catch (final IOException e) {
         LOGGER.error("Error writing document", e);
 
@@ -319,7 +555,10 @@ public class ChallengeDescriptionEditor extends JFrame {
     final JToolBar toolbar = new JToolBar();
     toolbar.setFloatable(false);
 
+    toolbar.add(mNewAction);
+    toolbar.add(mOpenAction);
     toolbar.add(mSaveAction);
+    toolbar.add(mSaveAsAction);
 
     return toolbar;
   }
